Progress save over
This was originally built on CodeCrafters, but
I decided I want to change up a few things that
the tests weren't flexible enough to still allow.
Unfortunately this means that my commit history
up until now is lost. I don't really know how
to also just save that over without all the
CodeCrafters extra junk that was packed in there.